+++ src/components/mod.rs
@@ -18,6 +18,8 @@ pub use commitlist::CommitList;
 pub use cred::CredComponent;
 pub use diff::DiffComponent;
 pub use revision_files::RevisionFilesComponent;
+use strum::EnumIter;
+use strum_macros::{EnumCount, EnumIs};
 pub use syntax_text::SyntaxTextComponent;
 pub use textinput::{InputType, TextInputComponent};
 pub use utils::{
@@ -189,6 +191,16 @@ pub enum EventState {
 pub enum FuzzyFinderTarget {
 	Branches,
 	Files,
+}
+
+#[derive(Copy, Clone, EnumCount, EnumIs, EnumIter)]
+pub enum BranchListSortBy {
+	BranchNameAsc,
+	BranchNameDesc,
+	LastCommitTimeDesc,
+	LastCommitTimeAsc,
+	LastCommitAuthorAsc,
+	LastCommitAuthorDesc,
 }
 
 impl EventState {